Add render_editor
