+++ src/engine/editor.h
@@ -2,6 +2,8 @@
 
 namespace platform {
 	struct Input;
+	class Renderer;
+	struct Canvas;
 }
 
 namespace engine {
@@ -13,5 +15,6 @@ namespace engine {
 
 	EditorState init_editor();
 	void update_editor(EditorState* editor, const platform::Input* input);
+	void render_editor(platform::Renderer* renderer, const EditorState* editor, platform::Canvas canvas);
 
 } // namespace engine
