+++ src/platform/texture.h
@@ -1,13 +1,13 @@
 #pragma once
 
 #include <SDL2/SDL_opengl.h>
+#include <glm/vec2.hpp>
 
 namespace platform {
 
 	struct Texture {
 		GLuint id;
-		int width;
-		int height;
+		glm::vec2 size;
 	};
 
 	Texture add_texture(const unsigned char* data, int width, int height);
